requirements.txt
pycryptodome==3.9.7
cffi==1.14.0
python-camellia==1.0